requirements.txt
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cobble==0.1.3
dnspython==2.2.1
docx==0.2.4
docx2pdf==0.1.8
dropbox==11.33.0
email-validator==1.2.1
et-xmlfile==1.1.0
Flask==2.1.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
forex-python==1.8
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
mammoth==1.4.15
MarkupSafe==2.1.1
openpyxl==3.0.10
Pillow==9.2.0
ply==3.11
pycparser==2.21
pypng==0.20220715.0
PyQRCode==1.2.1
python-dateutil==2.8.2
python-docx==0.8.11
python-slugify==6.1.2
requests==2.28.1
simplejson==3.17.6
six==1.16.0
SQLAlchemy==1.4.39
stone==3.3.1
text-unidecode==1.3
tqdm==4.64.0
urllib3==1.26.10
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.1
